// src/models/ITasks.ts
export interface ITasks {
    __metadata: {
        id: string;
        uri: string;
        etag: string;
        type: string;
    }
    ;
    Author:{
        __metadata: {
            id: string;
            type: string;
        }
        ID: number;
        Title: string;
    };
    Id: number;
    ID: number;
    Title: string;
    Created: string;
    Priority: string;
    Status: string;
    StartDate?: string;
    DueDate?: string;
  }